requirements.txt
pymongo==3.11.4
python-dotenv==0.17.1
fastapi[all]==0.65.1